backend/requirements.txt
# ═══════════════════════════════════════════════════════════════
# AI-SupraAgent Backend - Python Dependencies
# ═══════════════════════════════════════════════════════════════

# ─────────────────────────────────────────────────────────────────
# Web Framework
# ─────────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
sse-starlette>=2.0.0

# ─────────────────────────────────────────────────────────────────
# Validation & Settings
# ─────────────────────────────────────────────────────────────────
pydantic>=2.6.0
pydantic-settings>=2.2.0

# ─────────────────────────────────────────────────────────────────
# Database (PostgreSQL + pgvector)
# ─────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.25
asyncpg>=0.29.0
pgvector>=0.2.5
greenlet>=3.0.0

# ─────────────────────────────────────────────────────────────────
# Google Cloud / AI
# ─────────────────────────────────────────────────────────────────
google-cloud-aiplatform>=1.43.0
google-auth>=2.28.0
google-analytics-data>=0.18.0

# ─────────────────────────────────────────────────────────────────
# Document Processing
# ─────────────────────────────────────────────────────────────────
pypdf>=4.0.0
python-docx>=1.1.0

# ─────────────────────────────────────────────────────────────────
# MCP Protocol
# ─────────────────────────────────────────────────────────────────
mcp>=1.0.0

# ─────────────────────────────────────────────────────────────────
# Utilities
# ─────────────────────────────────────────────────────────────────
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
structlog>=24.1.0
tenacity>=8.2.0
aiofiles>=23.2.0

# ─────────────────────────────────────────────────────────────────
# Testing
# ─────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
